requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
websockets>=13.0
google-generativeai>=0.8.0
google-cloud-aiplatform>=1.70.0
google-cloud-firestore>=2.19.0
google-adk>=0.3.0
playwright>=1.49.0
pillow>=11.0.0
imagehash>=4.3.1
numpy>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.10.0
httpx>=0.28.0
mss>=9.0.0
pyautogui>=0.9.54
